fix(web): align API versioning with Next.js proxy
